refactor: 모달 코드를 page.tsx로 부터 분리, analysisService.ts lib/api 폴더 내부로 이동
